+++ CHANGELOG.md
@@ -2,6 +2,13 @@
 
 All notable changes to this project will be documented in this file.
 
+## [1.3.0] - 2023-08-17
+
+### Changed
+
+- Updated Lighthouse API to NOT load by default.
+- `PRODUCTION_DOMAIN` ENV Variable (versus no-reliable flag for production domain in API response)
+
 ## [1.2.0] - 2023-07-20
 
 ### Added